+++ src/mapfile_parser/internals/__init__.py
@@ -1,0 +1,13 @@
+#!/usr/bin/env python3
+
+# SPDX-FileCopyrightText: © 2025 Decompollaborate
+# SPDX-License-Identifier: MIT
+
+"""
+Do not use the definitions of this module outside of this library.
+Everything here may change at any time without prior notice.
+"""
+
+from __future__ import annotations
+
+from . import objdiff_report as objdiff_report